Added webhook event listener. (#850)
To donavanbecker,

If this PR doesn't agree to the strategy/policy of plug-in and/or target
user, reject it immediately. I guess there are any reasons for the
absence of this function.

To developers,

This PR implements webhook event listener to deliver them to the handler
of each device, as a basis of future development.

Config keyword 'webhookURL' in 'options' is added to specify webhook URL
which needs to be public server address where the plug-in is running.

~~~
{
  "name": "SwitchBot",
  "credentials": {...},
  "options": {
    "webhookURL": "http://${FQDN}:${PORT}/${PATH}",
  }
  ...
}
~~~

To implement the webhook event handler, look constructor() in meter.ts
as an example. The template code structure is as follows.

~~~
if (this.device.webhook) {
this.infoLog(`${this.device.deviceType}: ${this.accessory.displayName}
is listening webhook.`);
this.platform.webhookEventHandler[this.device.deviceId] = async
(context) => {
    try {
this.debugLog(`${this.device.deviceType}: ${this.accessory.displayName}
received Webhook: ${JSON.stringify(context)}`);
      // 
      // event processing in here.
      // 
    } catch (e: any) {
this.errorLog(`${this.device.deviceType}: ${this.accessory.displayName}
failed to handle webhook. Received: ${JSON.stringify(context)} Error:
${e}`);
    }
  }
}
~~~

and the device handler is enabled by config kwyword 'webhook' of device
in 'devices' section.

~~~
"devices":[
  {
    "deviceId": "...",
    "webhook": true,
    ...
  },
  ...	
]
~~~

I will add development tips, if approved.

Co-authored-by: Donavan Becker <beckersmarthome@icloud.com>