Feature/add latch to lock #833 (#859)
## :recycle: Current situation

Currently, the Homebridge plugin for the SwitchBot lock does not support
a "Latch" function. This means that users are unable to re-unlock the
door if it's already unlocked without first locking it.

## :bulb: Proposed solution

The proposed solution is to introduce a "Latch" function that acts as a
stateless trigger, sending an unlock command to the lock regardless of
its current state. This feature will act like a push button, where the
switch will return to "off" after activation, mimicking the behavior of
a physical latch button.

## :gear: Release Notes

The "Latch" feature is optional and can be enabled in the "SwitchBot
Device Settings".

### Testing

### Reviewer Nudging

---------

Co-authored-by: Dave Nicolson <david.nicolson@gmail.com>
Co-authored-by: Quentin Ulmer <quentin.ulmer@stud.h-da.de>
Co-authored-by: Donavan Becker <beckersmarthome@icloud.com>