Fix max retry (#1088)
## :recycle: Current situation

The value `0` cannot be used as it is evaluated in the same way as being
unset.

## :bulb: Proposed solution

Check the actual type of the value to allow 0 retries.

This is needed with v4.0.0+ as `READ_TIMEOUT` errors occur even on
success, resulting in erratic retry behavior. Using v.3.7.0 is still the
safest option.